ml_service/requirements.txt
# Core ML
tensorflow==2.15.0
numpy==1.26.4
scikit-learn==1.4.2
pandas==2.2.2
joblib==1.4.2

# Image processing
Pillow==10.3.0
opencv-python==4.9.0.80

# Data visualization (notebook)
matplotlib==3.8.4
seaborn==0.13.2

# Dataset download
kagglehub==0.2.7

# Excel support (clinical dataset)
openpyxl==3.1.2

# ML API service
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9

# Model optimization / mobile
tflite-runtime==2.14.0
